+++ src/lib/config.ts
@@ -61,9 +61,9 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 
 export const SDK_METADATA = {
   language: "typescript",
-  openapiDocVersion: "4.14.0-alpha.1755764858399-d5c9979b",
-  sdkVersion: "0.0.44-alpha.1",
+  openapiDocVersion: "4.14.0-alpha.1755863151363-73823764",
+  sdkVersion: "0.0.45-alpha.1",
   genVersion: "2.686.7",
   userAgent:
-    "speakeasy-sdk/typescript 0.0.44-alpha.1 2.686.7 4.14.0-alpha.1755764858399-d5c9979b cribl-control-plane",
+    "speakeasy-sdk/typescript 0.0.45-alpha.1 2.686.7 4.14.0-alpha.1755863151363-73823764 cribl-control-plane",
 } as const;